+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - LLM Judgement Customized Prompt Template Implementation  [#264](https://github.com/opensearch-project/search-relevance/pull/264)
 - Add `_search` endpoint for searching for Search Configurations using OpenSearch DSL  [#372](https://github.com/opensearch-project/search-relevance/pull/372)
 - Add `_search` endpoint for searching for Judgments using OpenSearch DSL  [#371](https://github.com/opensearch-project/search-relevance/pull/371)
+- Add `_search` endpoint for searching for Query Sets using OpenSearch DSL  [#362](https://github.com/opensearch-project/search-relevance/pull/362)
+
 
 ### Enhancements
 - Add BWC and Integration Tests for Index Mapping Update with Schema Version [#349](https://github.com/opensearch-project/search-relevance/pull/349)